requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.
djangorestframework-simplejwt==5.0.0
django-sslserver==0.22
python-dotenv==0.19.2
channels==3.0.4
